+++ wp-content/themes/ednc-2016/404.php
@@ -1,4 +1,4 @@
-<?php get_template_part('templates/page', 'header'); ?>
+<?php get_template_part('templates/components/page', 'header'); ?>
 
 <div class="alert alert-warning">
   <?php _e('Sorry, but the page you were trying to view does not exist.', 'sage'); ?>