+++ wp-content/themes/ednc-2016/base.php
@@ -7,7 +7,7 @@ use Roots\Sage\Wrapper;
 
 <!doctype html>
 <html <?php language_attributes(); ?>>
-  <?php get_template_part('templates/head'); ?>
+  <?php get_template_part('templates/layouts/head'); ?>
   <body <?php body_class(); ?>>
     <!--[if IE]>
       <div class="alert alert-warning">
@@ -16,9 +16,9 @@ use Roots\Sage\Wrapper;
     <![endif]-->
     <?php
       do_action('get_header');
-      get_template_part('templates/alert');
-      get_template_part('templates/header', 'sm-down');
-      get_template_part('templates/header', 'md-up');
+      get_template_part('templates/components/alert');
+      get_template_part('templates/layouts/header', 'sm-down');
+      get_template_part('templates/layouts/header', 'md-up');
     ?>
     <div class="wrap" role="document">
       <div class="content clearfix">
@@ -35,8 +35,8 @@ use Roots\Sage\Wrapper;
 
     <?php
       do_action('get_footer');
-      get_template_part('templates/footer');
-      get_template_part('templates/mobile-ad');
+      get_template_part('templates/layouts/footer');
+      get_template_part('templates/components/mobile-ad');
       wp_footer();
     ?>
   </body>