+++ wp-content/themes/ednc-2016/search.php
@@ -1,4 +1,4 @@
-<?php get_template_part('templates/page', 'header'); ?>
+<?php get_template_part('templates/components/page', 'header'); ?>
 
 <?php if (!have_posts()) : ?>
   <div class="alert alert-warning">
@@ -8,7 +8,7 @@
 <?php endif; ?>
 
 <?php while (have_posts()) : the_post(); ?>
-  <?php get_template_part('templates/content', 'search'); ?>
+  <?php get_template_part('templates/layouts/content', 'search'); ?>
 <?php endwhile; ?>
 
 <?php the_posts_navigation(); ?>