+++ wp-content/themes/ednc-2016/template-custom.php
@@ -5,6 +5,6 @@
 ?>
 
 <?php while (have_posts()) : the_post(); ?>
-  <?php get_template_part('templates/page', 'header'); ?>
-  <?php get_template_part('templates/content', 'page'); ?>
+  <?php get_template_part('templates/components/page', 'header'); ?>
+  <?php get_template_part('templates/layouts/content', 'page'); ?>
 <?php endwhile; ?>