+++ wp-content/themes/ednc-2016/templates/layouts/head.php
@@ -25,7 +25,7 @@
 
   <?php
   if (!is_user_logged_in()) {
-    get_template_part('templates/analytics');
+    get_template_part('templates/components/analytics');
   }
   ?>